Add missing Facts assignment in MessageCardSection
This was intended to go in with the other work for
dasrick/go-teams-notify#6, but I think this assignment
statement was unintentionally trimmed alongside the
removal of the package-level logger statements referenced
in dasrick/go-teams-notify#10.

I caught this when performing a test rebase to audit
what remaining changes in my fork haven't been included
or proposed in the parent/upstream project.